permissions done. login, logout, sessions done. cleaning up

+++ db/migrate/20131119000106_create_notes.rb
@@ -2,7 +2,7 @@ class CreateNotes < ActiveRecord::Migration
   def change
     create_table :notes do |t|
       t.text :body
-      t.integer :video_timestamp
+      t.float :video_timestamp
       t.integer :video_id
 
       t.timestamps
